chore: restructure css package
